web/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
pillow==11.0.0

